@@ -20,87 +19,54 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^10.0.8",
-    "@angular/core": "^10.0.8",
-    "@angular/fire": "^6.0.2",
-    "@angular/forms": "^10.0.8",
-    "@angular/platform-browser": "^10.0.8",
-    "@angular/platform-browser-dynamic": "^10.0.8",
-    "@angular/pwa": "^0.1000.5",
-    "@angular/router": "^10.0.8",
-    "@angular/service-worker": "^10.0.8",
-    "@capacitor/android": "^2.4.0",
-    "@capacitor/cli": "^2.4.0",
-    "@capacitor/core": "^2.4.0",
-    "@ionic-native/core": "5.27.0",
-    "@ionic-native/in-app-browser": "5.27.0",
-    "@ionic-native/splash-screen": "5.27.0",
-    "@ionic-native/status-bar": "5.27.0",
-    "@ionic/angular": "^5.3.1",
-    "@ionic/storage": "^2.3.0",
-    "angularfire2": "^5.4.2",
-    "cordova-android": "^9.0.0",
-    "cordova-ios": "^6.1.0",
-    "cordova-plugin-device": "^2.0.3",
-    "cordova-plugin-inappbrowser": "^4.0.0",
-    "cordova-plugin-ionic-keyboard": "^2.2.0",
-    "cordova-plugin-ionic-webview": "^5.0.0",
-    "cordova-plugin-splashscreen": "^6.0.0",
-    "cordova-plugin-statusbar": "^2.4.3",
-    "cordova-plugin-whitelist": "^1.3.4",
-    "core-js": "^3.6.5",
-    "firebase": "^7.17.2",
-    "leaflet": "^1.6.0",
+    "@angular/common": "^17.1.0",
+    "@angular/core": "^17.3.12",
+    "@angular/fire": "^17.1.0",
+    "@angular/forms": "^17.1.0",
+    "@angular/platform-browser": "^17.1.0",
+    "@angular/platform-browser-dynamic": "^17.1.0",
+    "@angular/pwa": "^17.1.0",
+    "@angular/router": "^17.1.0",
+    "@angular/service-worker": "^17.1.0",
+    "@capacitor/android": "^7.0.0",
+    "@capacitor/app": "^7.0.0",
+    "@capacitor/browser": "^7.0.0",
+    "@capacitor/cli": "^7.0.0",
+    "@capacitor/core": "^7.0.0",
+    "@capacitor/splash-screen": "^7.0.0",
+    "@capacitor/status-bar": "^7.0.0",
+    "@ionic/angular": "^7.1.1",
+    "@ionic/core": "^7.1.1",
+    "@ionic/pwa-elements": "^3.3.0",
+    "@ionic/storage": "^3.0.6",
+    "@ionic/storage-angular": "^4.0.0",
+    "firebase": "^10.14.1",
+    "leaflet": "^1.9.4",
     "leaflet-ant-path": "^1.3.0",
-    "rxjs": "^6.6.2",
-    "sw-toolbox": "3.6.0",
-    "tslib": "^2.0.1",
-    "zone.js": "~0.10.3"
+    "rxjs": "^7.8.1",
+    "tslib": "^2.8.1",
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1000.5",
-    "@angular-devkit/build-angular": "^0.1000.5",
-    "@angular-devkit/core": "^10.0.5",
-    "@angular-devkit/schematics": "^10.0.5",
-    "@angular/cli": "^10.0.5",
-    "@angular/compiler": "^10.0.8",
-    "@angular/compiler-cli": "^10.0.8",
-    "@angular/language-service": "^10.0.8",
-    "@ionic/angular-toolkit": "^2.3.0",
-    "@ionic/lab": "^3.1.7",
-    "@types/jasmine": "3.5.12",
+    "@angular-devkit/architect": "^0.1701.0",
+    "@angular-devkit/build-angular": "^17.1.0",
+    "@angular-devkit/core": "^17.1.0",
+    "@angular-devkit/schematics": "^17.1.0",
+    "@angular/cli": "^17.1.0",
+    "@angular/compiler": "^17.1.0",
+    "@angular/compiler-cli": "^17.1.0",
+    "@angular/language-service": "^17.1.0",
+    "@ionic/angular-toolkit": "^8.0.0",
+    "@types/jasmine": "~4.3.0",
     "@types/jasminewd2": "2.0.8",
-    "@types/node": "^14.0.27",
-    "codelyzer": "^6.0.0",
-    "cordova-res": "^0.15.1",
-    "jasmine-core": "^3.6.0",
+    "@types/node": "^20.10.0",
+    "jasmine-core": "~4.6.0",
     "jasmine-spec-reporter": "5.0.2",
-    "karma": "^5.1.1",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "3.3.1",
-    "karma-jasmine-html-reporter": "1.5.4",
-    "protractor": "^7.0.0",
-    "puppeteer": "^5.2.1",
-    "ts-node": "8.10.2",
-    "tslint": "^6.1.3",
-    "typescript": "~3.9.7"
-  },
-  "cordova": {
-    "platforms": [
-      "ios",
-      "android"
-    ],
-    "plugins": {
-      "cordova-plugin-device": {},
-      "cordova-plugin-inappbrowser": {},
-      "cordova-plugin-splashscreen": {},
-      "cordova-plugin-whitelist": {},
-      "cordova-plugin-ionic-webview": {
-        "ANDROID_SUPPORT_ANNOTATIONS_VERSION": "27.+"
-      },
-      "cordova-plugin-statusbar": {},
-      "cordova-plugin-ionic-keyboard": {}
-    }
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.3.2"
   }
 }
