@@ -1,11 +1,14 @@
-import { Injectable } from '@angular/core';
 import { CanLoad, Router } from '@angular/router';
-import { Storage } from '@ionic/storage';
+import { Injectable, inject } from '@angular/core';
+
+import { Storage } from '@ionic/storage-angular';
+
 @Injectable({
   providedIn: 'root'
 })
 export class CheckTutorial implements CanLoad {
-  constructor(private storage: Storage, private router: Router) {}
+  private storage = inject(Storage);
+  private router = inject(Router);
 
   canLoad() {
     return this.storage.get('ion_did_tutorial').then(res => {
