@@ -1,27 +1,26 @@
-import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { InAppBrowser } from '@ionic-native/in-app-browser/ngx';
-import { SplashScreen } from '@ionic-native/splash-screen/ngx';
-import { StatusBar } from '@ionic-native/status-bar/ngx';
-import { IonicModule } from '@ionic/angular';
-import { IonicStorageModule } from '@ionic/storage';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
+import { HttpClientModule } from '@angular/common/http';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { ServiceWorkerModule } from '@angular/service-worker';
-import { environment } from '../environments/environment';
-import { FormsModule } from '@angular/forms';
+import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
+import { IonicStorageModule } from '@ionic/storage-angular';
+import { RouteReuseStrategy } from '@angular/router';
 
-//firebase
-import { AngularFireModule } from '@angular/fire';
-//import { AngularFireAuthModule } from '@angular/fire/auth';
-import { AngularFireDatabaseModule } from '@angular/fire/database';
-//import { AngularFireStorageModule } from '@angular/fire/storage';
+// App imports
+import { AppRoutingModule } from './app-routing.module';
 import { AuthenticationService } from './services/authentication.service';
-import * as firebase from 'firebase';
-import { ReactiveFormsModule } from '@angular/forms';
+import { environment } from '../environments/environment';
 
+// Import Firebase manually to avoid the providers issue
+import { initializeApp } from 'firebase/app';
+import { getAuth } from 'firebase/auth';
+import { getDatabase } from 'firebase/database';
+
+// Initialize Firebase directly
+const app = initializeApp(environment.firebaseConfig);
+const auth = getAuth(app);
+const database = getDatabase(app);
 
 @NgModule({
   imports: [
@@ -30,16 +29,23 @@ import { ReactiveFormsModule } from '@angular/forms';
     HttpClientModule,
     FormsModule,
     ReactiveFormsModule,
-    AngularFireModule.initializeApp(environment.firebaseConfig),
-    AngularFireDatabaseModule,
+
+    // Ionic
     IonicModule.forRoot(),
     IonicStorageModule.forRoot(),
+
+    // PWA
     ServiceWorkerModule.register('ngsw-worker.js', {
-      enabled: environment.production
+      enabled: environment.production,
+      registrationStrategy: 'registerWhenStable:30000'
     })
   ],
-  declarations: [AppComponent],
-  providers: [InAppBrowser, SplashScreen, StatusBar],
-  bootstrap: [AppComponent]
+  providers: [
+    AuthenticationService,
+    { provide: 'app', useValue: app },
+    { provide: 'auth', useValue: auth },
+    { provide: 'database', useValue: database },
+    { provide: RouteReuseStrategy, useClass: IonicRouteStrategy }
+  ]
 })
 export class AppModule {}
