@@ -1,4 +1,4 @@
-import { NgModule } from '@angular/core';
+import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
@@ -15,7 +15,7 @@ import { AboutPageRoutingModule } from './about-routing.module';
     AboutPageRoutingModule
   ],
   declarations: [AboutPage, PopoverPage],
-  entryComponents: [PopoverPage],
   bootstrap: [AboutPage],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class AboutModule {}
