@@ -1,6 +1,6 @@
-import { NgModule } from '@angular/core';
+import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 
 import { SchedulePage } from './schedule';
@@ -11,6 +11,7 @@ import { SchedulePageRoutingModule } from './schedule-routing.module';
   imports: [
     CommonModule,
     FormsModule,
+    ReactiveFormsModule,
     IonicModule,
     SchedulePageRoutingModule
   ],
@@ -18,8 +19,6 @@ import { SchedulePageRoutingModule } from './schedule-routing.module';
     SchedulePage,
     ScheduleFilterPage
   ],
-  entryComponents: [
-    ScheduleFilterPage
-  ]
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class ScheduleModule { }
