@@ -1,8 +1,8 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-import { IonicModule } from '@ionic/angular';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
+import { CommonModule } from '@angular/common';
+import { IonicModule } from '@ionic/angular';
 import { SignupPage } from './signup';
 import { SignupPageRoutingModule } from './signup-routing.module';
 
@@ -10,11 +10,13 @@ import { SignupPageRoutingModule } from './signup-routing.module';
   imports: [
     CommonModule,
     FormsModule,
+    ReactiveFormsModule,
     IonicModule,
     SignupPageRoutingModule
   ],
   declarations: [
     SignupPage,
-  ]
+  ],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class SignUpModule { }
