@@ -1,18 +1,11 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
 import { SpeakerDetailPage } from './speaker-detail';
 import { SpeakerDetailPageRoutingModule } from './speaker-detail-routing.module';
-import { IonicModule } from '@ionic/angular';
 
 @NgModule({
   imports: [
-    CommonModule,
-    IonicModule,
-    SpeakerDetailPageRoutingModule
-  ],
-  declarations: [
     SpeakerDetailPage,
+    SpeakerDetailPageRoutingModule
   ]
 })
 export class SpeakerDetailModule { }
