@@ -1,7 +1,8 @@
-import { NgModule } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
+
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
-
+import { IonicStorageModule } from '@ionic/storage-angular';
 import { TutorialPage } from './tutorial';
 import { TutorialPageRoutingModule } from './tutorial-routing.module';
 
@@ -9,9 +10,10 @@ import { TutorialPageRoutingModule } from './tutorial-routing.module';
   imports: [
     CommonModule,
     IonicModule,
+    IonicStorageModule.forRoot(),
     TutorialPageRoutingModule
   ],
   declarations: [TutorialPage],
-  entryComponents: [TutorialPage],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class TutorialModule {}
