@@ -1,9 +1,8 @@
 import { Component, ViewChild } from '@angular/core';
+import { IonicSlides, MenuController } from '@ionic/angular';
+
 import { Router } from '@angular/router';
-
-import { MenuController, IonSlides } from '@ionic/angular';
-
-import { Storage } from '@ionic/storage';
+import { Storage } from '@ionic/storage-angular';
 
 @Component({
   selector: 'page-tutorial',
@@ -13,13 +12,19 @@ import { Storage } from '@ionic/storage';
 export class TutorialPage {
   showSkip = true;
 
-  @ViewChild('slides', { static: true }) slides: IonSlides;
+  @ViewChild('slides') slides: any;
 
   constructor(
     public menu: MenuController,
     public router: Router,
     public storage: Storage
-  ) {}
+  ) {
+    this.initStorage();
+  }
+
+  async initStorage() {
+    await this.storage.create();
+  }
 
   startApp() {
     this.router
