@@ -1,6 +1,6 @@
-import { NgModule } from '@angular/core';
+import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { IonicModule } from '@ionic/angular';
 
@@ -12,9 +12,11 @@ import { UserReportsPage } from './user-reports.page';
   imports: [
     CommonModule,
     FormsModule,
+    ReactiveFormsModule,
     IonicModule,
     UserReportsPageRoutingModule
   ],
-  declarations: [UserReportsPage]
+  declarations: [UserReportsPage],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class UserReportsPageModule {}
