@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Storage } from '@ionic/storage';
-
+import { Storage } from '@ionic/storage-angular';
 
 @Injectable({
   providedIn: 'root'
