@@ -1,64 +1,66 @@
-import * as firebase from 'firebase/app';
-
-import { AngularFireAuthModule } from '@angular/fire/auth';
-import { Injectable } from '@angular/core';
+import { Injectable, Inject } from '@angular/core';
+import {
+  Auth,
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
+  signOut,
+  onAuthStateChanged,
+  sendPasswordResetEmail,
+  User
+} from 'firebase/auth';
+import { Observable, from, BehaviorSubject } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
 })
-
 export class AuthenticationService {
+  private currentUserSubject = new BehaviorSubject<User | null>(null);
+  currentUser$ = this.currentUserSubject.asObservable();
 
-  constructor($scope: any) {
-    var user = firebase.auth().currentUser;
-    if(user != null){
-
-    }
+  constructor(@Inject('auth') private auth: Auth) {
+    // Initialize the current user from the auth state
+    onAuthStateChanged(this.auth, (user) => {
+      this.currentUserSubject.next(user);
+    });
   }
 
-  registerUser(value){
-   return new Promise<any>((resolve, reject) => {
-     firebase.auth().createUserWithEmailAndPassword(value.email, value.password)
-     .then(
-       res => resolve(res),
-       err => reject(err))
-   })
+  async isAuthenticated(): Promise<boolean> {
+    return new Promise((resolve) => {
+      onAuthStateChanged(this.auth, (user) => {
+        resolve(!!user);
+      });
+    });
   }
 
-  loginUser(value){
-   return new Promise<any>((resolve, reject) => {
-     firebase.auth().signInWithEmailAndPassword(value.email, value.password)
-     .then(
-       res => resolve(res),
-       err => reject(err))
-   })
+  registerUser(value: { email: string, password: string }) {
+    return from(createUserWithEmailAndPassword(
+      this.auth,
+      value.email,
+      value.password
+    ));
   }
 
-  logoutUser(){
-    return new Promise<void>((resolve, reject) => {
-      if(firebase.auth().currentUser){
-        firebase.auth().signOut()
-        .then(() => {
-          console.log("LOG Out");
-          resolve();
-        }).catch((error) => {
-          reject();
-        });
-      }
-    })
+  login(value: { email: string, password: string }) {
+    return from(signInWithEmailAndPassword(
+      this.auth,
+      value.email,
+      value.password
+    ));
   }
 
-  userDetails(){
-    firebase.auth().onAuthStateChanged((user) => {
-      console.log(user)
-    })
-    // firebase.auth().currentUser;
+  logoutUser() {
+    return from(signOut(this.auth));
   }
-  currentUserWithScope(){
-    var user = firebase.auth().currentUser;
-    if(user != null){
 
-    }
-    return user;
+  userDetails() {
+    return this.currentUser$;
+  }
+
+  currentUserWithScope() {
+    return this.auth.currentUser;
+  }
+
+  resetPassword(email: string) {
+    return from(sendPasswordResetEmail(this.auth, email));
   }
 }
