@@ -1,3 +1,15 @@
 export const environment = {
-  production: true
+  production: true,
+  appName: 'Safe Space',
+  version: '1.0.0',
+  firebaseConfig: {
+    apiKey: "AIzaSyCcgIFxX8VE8cONPSCvE9ylhhAaK1sT4bs",
+    authDomain: "ionic-92142.firebaseapp.com",
+    databaseURL: "https://ionic-92142.firebaseio.com",
+    projectId: "ionic-92142",
+    storageBucket: "ionic-92142.appspot.com",
+    messagingSenderId: "214578467644",
+    appId: "1:214578467644:web:2c95d6bb23051523540439",
+    measurementId: "G-3BEDB2LTV1"
+  }
 };
