@@ -1,9 +1,11 @@
 // The file contents for the current environment will overwrite these during build.
 // The build system defaults to the dev environment which uses `environment.ts`, but if you do
-// `ng build --env=prod` then `environment.prod.ts` will be used instead.
+// `ng build --configuration=production` then `environment.prod.ts` will be used instead.
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 export const environment = {
   production: false,
+  appName: 'Safe Space',
+  version: '1.0.0',
   firebaseConfig: {
     apiKey: "AIzaSyCcgIFxX8VE8cONPSCvE9ylhhAaK1sT4bs",
     authDomain: "ionic-92142.firebaseapp.com",
@@ -13,8 +15,12 @@ export const environment = {
     messagingSenderId: "214578467644",
     appId: "1:214578467644:web:2c95d6bb23051523540439",
     measurementId: "G-3BEDB2LTV1"
-
-  }
+  },
+  // Additional configuration options
+  defaultLanguage: 'en',
+  supportedLanguages: ['en'],
+  apiTimeout: 30000,
+  enableAnalytics: false
 };
 
 /*
