@@ -1,12 +1,22 @@
+import { bootstrapApplication } from '@angular/platform-browser';
+import { AppComponent } from './app/app.component';
+import { appConfig } from './app/app.config';
 import { enableProdMode } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
-import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
+
+// Import Ionic web components
+import { defineCustomElements } from '@ionic/pwa-elements/loader';
+import { defineCustomElements as defineIonicCore } from '@ionic/core/loader';
 
 if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic().bootstrapModule(AppModule)
-  .catch(err => console.log(err));
+// Bootstrap Angular
+bootstrapApplication(AppComponent, appConfig)
+  .then(() => {
+    // Load Ionic web components after Angular is bootstrapped
+    defineCustomElements(window);
+    defineIonicCore(window);
+  })
+  .catch(err => console.error(err));
